add better csv conversion
converted csv files were causing issues with password managers, because of how csv was structured.
such issues resolved by implementing a better csv conversion algorithm.
now, while converting to csv, the text to be used in a row is matched against \' \, " symbols.
if comma is not present, text is preserved as is.
if comma is present, \' and " characters get counted. The one with 1 or less count is choosen to be wrapper around text.